Added SOC calcualation changes
Unusual case, have a leaf BMS with the current sensor put in the wrong way around. So SOC from BMS is rather off.
However it seems to update the SOC based on voltage when first starting up.
So I'll use that as a starting point and have zombie calculate SOC from that point.