Add monotonic "uptime" spot-value
It is useful to have a monotonic uptime counter to know how long
the device has been running and to spot unexpected reboots.

The uptime value is also exposed in the time an error is logged.
Having the current values allows it to be compared against the
time an error occurred.

Tests:
 - Verify that uptime updates in the web interface and in
   OpenInverter CAN Tool